Refactor code: create profile for login
